+++ app/lib/config/app_config.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/foundation.dart';
+
+@immutable
+final class AppConfig {
+  const AppConfig({this.isIntegrationTest = false});
+
+  final bool isIntegrationTest;
+}
