+++ app/lib/config/config.dart
@@ -1,0 +1,1 @@
+export 'app_config.dart';
