+++ app/pubspec.yaml
@@ -22,7 +22,6 @@ dependencies:
   dartz: ^0.10.1
   connectivity_plus: ^4.0.1
   http: ^1.0.0
-  web_socket_channel: ^2.4.0
   stomp_dart_client: ^0.4.4
   infinite_scroll_pagination: ^4.0.0-dev.1
   flutter_svg: ^2.0.7
@@ -30,6 +29,7 @@ dependencies:
   json_annotation: ^4.8.1
   google_fonts: ^5.0.0
   font_awesome_flutter: ^10.4.0
+  upgrader: ^7.1.0
   mq_storage:
     path: ../packages/mq_storage/
   # flutter_native_splash: ^2.2.18
