+++ app/test_driver/app.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:flutter_driver/driver_extension.dart';
+import 'package:my_quran/config/config.dart';
 import 'package:my_quran/main.dart' as app;
 
 // flutter drive --target=test_driver/app.dart
@@ -16,5 +17,5 @@ Future<void> main() async {
       return Future.value(result);
     },
   );
-  await app.main();
+  await app.main(appConfig: const AppConfig(isIntegrationTest: true));
 }
